restructure and reorganize mcmc scripts
